Fixed Issue #5
Issue #5 was solved, which was being caused by the MatrixBot incorrectly indexing the possible moves.

+++ Player/Bot/MatrixBot/MatrixBot.h
@@ -9,8 +9,8 @@ class MatrixBot : public Bot {
         std::vector<std::vector<float>> transition_matrix;
         std::vector<float> get_input_vector();
         std::vector<float> get_output_vector();
-        move get_move_from_index(int index);
-        Piece* get_piece_from_index(int index);
+        move get_move_from_index(long unsigned int index);
+        Piece* get_piece_from_index(long unsigned int index);
         
         public:
         MatrixBot(int index_i);
